testnet: Handle preloading using button in wallet panel
